requirements.txt
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
sentence-transformers==2.5.1
huggingface-hub==0.21.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
sqlalchemy==2.0.23
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
joblib==1.3.2
scipy==1.11.4